+++ test/util/pull-request-body.ts
@@ -90,6 +90,20 @@ describe('PullRequestBody', () => {
       expect(releaseData[0].version?.toString()).to.eql('0.3.0');
       expect(releaseData[0].notes).matches(/Database operations/);
     });
+
+    it('can parse initial release pull rqeuest body', () => {
+      const body = readFileSync(
+        resolve(fixturesPath, './initial-version.txt'),
+        'utf8'
+      );
+      const pullRequestBody = PullRequestBody.parse(body);
+      expect(pullRequestBody).to.not.be.undefined;
+      const releaseData = pullRequestBody!.releaseData;
+      expect(releaseData).lengthOf(1);
+      expect(releaseData[0].component).to.be.undefined;
+      expect(releaseData[0].version?.toString()).to.eql('0.1.0');
+      expect(releaseData[0].notes).matches(/initial generation/);
+    });
   });
   describe('toString', () => {
     it('can handle multiple entries', () => {
@@ -118,6 +132,18 @@ describe('PullRequestBody', () => {
         },
       ];
       const pullRequestBody = new PullRequestBody(data);
+      snapshot(pullRequestBody.toString());
+    });
+
+    it('can handle a single entries forced components', () => {
+      const data = [
+        {
+          component: 'pkg1',
+          version: Version.parse('1.2.3'),
+          notes: 'some special notes go here',
+        },
+      ];
+      const pullRequestBody = new PullRequestBody(data, {useComponents: true});
       snapshot(pullRequestBody.toString());
     });